requirements.txt
flet==0.27.6
SQLAlchemy==2.0.39
pymysql==1.1.1
python-dotenv==1.0.1


# para inatalar o requirements:
# deactivate
#   pip install -r requirements.txt


#   pip install flet
#   pip install sqlalchemy
#   pip install pymysql
#   pip install python-dotenv


# para rodar o codigo:
#   python src/main.py

#na web:
#flet run src/main.py --web
#flet src/main.py --web  

#no celular:
#flet run src/main.py --android
#flet src/main.py --android


#oque falta:

# scroll
# margem
# tabela
# alterar tema
# pagina responsiva
# simulador movel

